Updated PortfolioProject1
Queries used for Tableau Project
There are SEVERAL changes from the guided project mostly due to two 
distinct factors. First, he does not format his data in a way that works
with calculations, so we need to as in AS FLOAT or similar. 

Second, because the data is not clean (location and continent, as well as additional categories like income), 
two of the calculations are incorrect if you use his default code (by a lot, literally 5x as many cases worldwide).
The following code works with the data as it comes from https://ourworldindata.org/covid-deaths as of April 11, 2023